requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cssselect2==0.6.0
Django==4.0.5
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-qr-code==3.0.0
django-weasyprint==2.1.0
django-widget-tweaks==1.4.12
fonttools==4.33.3
html5lib==1.1
idna==3.3
jsonfield==3.1.0
pdfkit==1.0.0
Pillow==9.1.1
pycparser==2.21
pydyf==0.2.0
pyphen==0.12.0
pytz==2022.1
requests==2.27.1
segno==1.5.2
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
urllib3==1.26.9
webencodings==0.5.1
zopfli==0.2.1
